+++ azure_sdk_cosmos/src/document.rs
@@ -1,5 +1,5 @@
+use crate::from_headers::{number_of_read_regions_from_headers, request_charge_from_headers};
 use crate::DocumentAttributes;
-use crate::{number_of_read_regions_from_headers, request_charge_from_headers};
 use azure_sdk_core::errors::AzureError;
 use azure_sdk_core::session_token_from_headers;
 use hyper::header::HeaderMap;
