+++ azure_sdk_cosmos/src/responses/create_stored_procedure_response.rs
@@ -1,10 +1,6 @@
+use crate::from_headers::*;
 use crate::stored_procedure::StoredProcedure;
-use crate::{
-    activity_id_from_headers, current_replica_set_size_from_headers,
-    current_write_quorum_from_headers, last_state_change_from_headers,
-    quorum_hacked_lsn_from_headers, request_charge_from_headers, resource_quota_from_headers,
-    resource_usage_from_headers, ResourceQuota,
-};
+use crate::ResourceQuota;
 use azure_sdk_core::errors::AzureError;
 use azure_sdk_core::etag_from_headers;
 use azure_sdk_core::session_token_from_headers;
