+++ azure_sdk_cosmos/src/responses/delete_permission_response.rs
@@ -1,7 +1,4 @@
-use crate::{
-    activity_id_from_headers, alt_content_path_from_headers, content_path_from_headers,
-    request_charge_from_headers,
-};
+use crate::from_headers::*;
 use azure_sdk_core::errors::AzureError;
 use azure_sdk_core::session_token_from_headers;
 use http::HeaderMap;
