+++ azure_sdk_cosmos/src/responses/delete_stored_procedure_response.rs
@@ -1,7 +1,5 @@
-use crate::{
-    activity_id_from_headers, last_state_change_from_headers, request_charge_from_headers,
-    resource_quota_from_headers, resource_usage_from_headers, ResourceQuota,
-};
+use crate::from_headers::*;
+use crate::ResourceQuota;
 use azure_sdk_core::errors::AzureError;
 use azure_sdk_core::session_token_from_headers;
 use chrono::{DateTime, Utc};
