+++ azure_sdk_cosmos/src/responses/delete_user_response.rs
@@ -1,4 +1,4 @@
-use crate::{activity_id_from_headers, request_charge_from_headers};
+use crate::from_headers::*;
 use azure_sdk_core::errors::AzureError;
 use http::HeaderMap;
