+++ azure_sdk_cosmos/src/responses/execute_stored_procedure_response.rs
@@ -1,8 +1,4 @@
-use crate::{
-    activity_id_from_headers, current_replica_set_size_from_headers,
-    current_write_quorum_from_headers, last_state_change_from_headers,
-    quorum_hacked_lsn_from_headers, request_charge_from_headers,
-};
+use crate::from_headers::*;
 use azure_sdk_core::errors::AzureError;
 use azure_sdk_core::session_token_from_headers;
 use chrono::{DateTime, Utc};
