+++ azure_sdk_cosmos/src/responses/get_permission_response.rs
@@ -1,9 +1,6 @@
+use crate::from_headers::*;
 use crate::permission::CosmosPermission;
 use crate::Permission;
-use crate::{
-    activity_id_from_headers, alt_content_path_from_headers, content_path_from_headers,
-    request_charge_from_headers,
-};
 use azure_sdk_core::errors::AzureError;
 use azure_sdk_core::{etag_from_headers, session_token_from_headers};
 use http::HeaderMap;
