+++ azure_sdk_cosmos/src/responses/list_stored_procedures_response.rs
@@ -1,8 +1,6 @@
+use crate::from_headers::*;
 use crate::stored_procedure::StoredProcedure;
-use crate::{
-    activity_id_from_headers, last_state_change_from_headers, request_charge_from_headers,
-    resource_quota_from_headers, resource_usage_from_headers, ResourceQuota,
-};
+use crate::ResourceQuota;
 use azure_sdk_core::errors::AzureError;
 use azure_sdk_core::session_token_from_headers;
 use chrono::{DateTime, Utc};
