+++ azure_sdk_cosmos/src/responses/list_users_response.rs
@@ -1,5 +1,5 @@
+use crate::from_headers::*;
 use crate::User;
-use crate::{activity_id_from_headers, request_charge_from_headers};
 use azure_sdk_core::errors::AzureError;
 use http::HeaderMap;
 use serde::Deserialize;
